refactor: only generate node_modules on js templates
