+++ src/main/stubs/fxmanifestTemplate.ts
@@ -1,27 +1,29 @@
-export const createFxmaniest = (project: any) => {
-    if (project.template == 'TypeScript') {
+import { ProjectObject } from '../types/project';
+
+export const createFxmaniest = (project: ProjectObject) => {
+    if (project.resourceTemplate == 'TypeScript') {
         return `
 fx_version "cerulean"
 game "gta5"
-			
-author "${project.author}"
-version "${project.version}"
-description "${project.description}"
-			
+            
+author "${project.resourceAuthor}"
+version "${project.resourceVersion}"
+description "${project.resourceDescription}"
+            
 client_script "dist/client/*.client.js"
 server_script "dist/server/*.server.js"
-`;
+            `;
     } else {
         return `
 fx_version "cerulean"
 game "gta5"
-			
-author "${project.author}"
-version "${project.version}"
-description "${project.description}"
-			
-client_script "client/client.js"
-server_script "server/server.js"
-		`;
+            
+author "${project.resourceAuthor}"
+version "${project.resourceVersion}"
+description "${project.resourceDescription}"
+            
+client_script "client/*.lua"
+server_script "server/*.lua"
+        `;
     }
 };
